Rename mapping_{2,3}d::SparsePoseGraph.
[RFC=0001](https://github.com/googlecartographer/rfcs/blob/master/text/0001-renaming-sparse-pose-graph.md)
